chore: add api folder in app
