+++ src/app/flask_postgresql/api/event_handlers/event_handler.py
@@ -1,27 +1,30 @@
 """ This module implements the event handler for text message events.
 """
+from abc import abstractmethod
 from typing import Dict
+from linebot.v3.webhooks import MessageEvent
+
 from src.app.flask_postgresql.configs import Config
 from src.app.flask_postgresql.presenters.window_presenter import WindowPresenter
 from src.app.flask_postgresql.interfaces.event_handler_interface import EventHandlerInterface
-from src.app.flask_postgresql.controllers.text_message_content_controller import TextMessageContentController
 from src.infrastructure.repositories.window.window_postgresql_repository import WindowPostgresqlRepository
 from src.interactor.use_cases.create_window import CreateWindowUseCase
 from src.interactor.dtos.window_dtos import CreateWindowInputDto, GetWindowInputDto
 from src.interactor.use_cases.get_window import GetWindowUseCase
 from src.interactor.dtos.event_dto import EventInputDto
-from src.interactor.use_cases.create_text_message_reply import CreateTextMessageReplyUseCase
+from src.infrastructure.repositories.agent_chain.agent_chain_in_memory_repository import AgentExecutorInMemoryRepository
+from src.interactor.interfaces.logger.logger import LoggerInterface
 
 
-from linebot.v3.webhooks import MessageEvent
 
 
-class TextEventHandler(EventHandlerInterface):
-    def __init__(self, logger, agent_repository):
+class EventHandler(EventHandlerInterface):
+    def __init__(self, logger:LoggerInterface, agent_repository: AgentExecutorInMemoryRepository):
         self.logger = logger
         self.agent_repository = agent_repository
         self.input_dto: EventInputDto
 
+    
     def get_event_info(self, event: MessageEvent):
 
         if event.source.type == "user":
@@ -55,21 +58,7 @@ class TextEventHandler(EventHandlerInterface):
             window = self._create_window_info(window_id=window_id)
 
         return window
-   
-
-    def execute(self):
-
-        if self.input_dto.window.get("is_muting"):
-            return "靜悄悄的，什麼都沒有發生。"
-
-        use_case = CreateTextMessageReplyUseCase(
-            repository=self.agent_repository,
-            logger=self.logger,
-        )
-        result = use_case.execute(self.input_dto)
-        return result
-        
-
+    
     def _get_window_info(self, window_id: str):
         """
         Retrieves the information of a window.
@@ -86,7 +75,7 @@ class TextEventHandler(EventHandlerInterface):
         get_window_input_dto = GetWindowInputDto(window_id)
         result = use_case.execute(get_window_input_dto)
         return result
-
+    
     def _create_window_info(self, window_id: str):
         """
         Creates a new window info.
@@ -110,3 +99,10 @@ class TextEventHandler(EventHandlerInterface):
         )
         result = use_case.execute(create_window_input_dto)
         return result
+    
+    @abstractmethod
+    def execute(self):
+        """
+        A description of the entire function, its parameters, and its return types.
+        """
+
