+++ src/app/flask_postgresql/api/request_context.py
@@ -9,14 +9,14 @@ def setup_request_context(app: Flask):
     @app.before_request
     def before_request():
         """ Before Request """
-        print("before_request")
+        # print("before_request")
         g.db = Session()
 
     @app.teardown_request
     def teardown_request(_unused=False):
         """ After Request """
         dbc: scoped_session = getattr(g, 'db', None)
-        print("teardown_request")
+        # print("teardown_request")
         if dbc is not None:
             dbc.close()
